+++ src/Admin/views/boxzilla-nudge.php
@@ -1,0 +1,7 @@
+<?php defined( 'ABSPATH' ) or exit; ?>
+
+<div class="notice notice-warning">
+    <p style="font-weight: bold;">Scroll Triggered Boxes is now Boxzilla!</p>
+    <p>The <a href="https://boxzillaplugin.com/blog/2016/hello-boxzilla/">new Boxzilla plugin</a> comes with many improvements over Scroll Triggered Boxes, but also some breaking changes.</p>
+    <p>Whenever you're ready to make the switch, please <a href="https://kb.boxzillaplugin.com/updating-from-scroll-triggered-boxes/">read through the upgrade guide</a> for a full list of changes.</p>
+</div>